basic ds-submit and ds-retract. The dssubmitcommand and dsretractcommand are not yet implemented.
