+++ enforcer-ng/src/Makefile.am
@@ -121,11 +121,9 @@ ods_enforcerd_SOURCES = \
 	keystate/rollover_list_cmd.c keystate/rollover_list_cmd.h \
 	keystate/keystate_export_cmd.c keystate/keystate_export_cmd.h \
 	keystate/keystate_export_task.c keystate/keystate_export_task.h \
-	keystate/keystate_ds_submit_cmd.cpp keystate/keystate_ds_submit_cmd.h \
-	keystate/keystate_ds_submit_task.cpp keystate/keystate_ds_submit_task.h \
+	keystate/keystate_ds_submit_cmd.c keystate/keystate_ds_submit_cmd.h \
 	keystate/keystate_ds_seen_cmd.c keystate/keystate_ds_seen_cmd.h \
-	keystate/keystate_ds_retract_cmd.cpp keystate/keystate_ds_retract_cmd.h \
-	keystate/keystate_ds_retract_task.cpp keystate/keystate_ds_retract_task.h \
+	keystate/keystate_ds_retract_cmd.c keystate/keystate_ds_retract_cmd.h \
 	keystate/keystate_ds.c keystate/keystate_ds.h \
 	keystate/keystate_ds_gone_cmd.c keystate/keystate_ds_gone_cmd.h \
 	keystate/keystate_rollover_cmd.c keystate/keystate_rollover_cmd.h \
