+++ enforcer-ng/src/daemon/cmdhandler.c
@@ -133,10 +133,10 @@ cmd_funcs_avail(void)
 
         &key_list_funcblock,
         &key_export_funcblock,
-/*        &key_ds_submit_funcblock,
-*/        &key_ds_seen_funcblock,
-/*        &key_ds_retract_funcblock,
-*/        &key_ds_gone_funcblock,
+        &key_ds_submit_funcblock,
+        &key_ds_seen_funcblock,
+        &key_ds_retract_funcblock,
+        &key_ds_gone_funcblock,
         &key_generate_funcblock,
 
         &key_rollover_funcblock,
