+++ enforcer-ng/src/keystate/keystate_ds_submit_cmd.c
@@ -31,15 +31,14 @@
 
 #include "daemon/cmdhandler.h"
 #include "daemon/engine.h"
-#include "keystate/keystate_ds_submit_task.h"
-#include "shared/file.h"
+#include "enforcer/enforce_task.h"
 #include "shared/log.h"
 #include "shared/str.h"
 #include "daemon/clientpipe.h"
+#include "db/key_data.h"
+#include "keystate/keystate_ds.h"
 
 #include "keystate/keystate_ds_submit_cmd.h"
-
-static const char *module_str = "keystate_ds_submit_cmd";
 
 static void
 usage(int sockfd)
@@ -47,13 +46,10 @@ usage(int sockfd)
 	client_printf(sockfd,
 		"key ds-submit          Issue a ds-submit to the enforcer for a KSK.\n"
 		"                       (This command with no parameters lists eligible keys.)\n"
-		"      [--cka_id <CKA_ID>]        (aka -k)  cka_id <CKA_ID> of the key.\n"			
-		"      [--zone <zone> | --auto]   (aka -z | -a) specify a zone to submit keys\n"			
-		"                                           for or perform auto submit for all\n"
-		"                                           keys on all zones that have the\n"
-		"                                           submit flag set.\n"
-		"      [--force]                  (aka -f)  force even if there is no configured\n"
-		"                                           DelegationSignerSubmitCommand.\n"
+		"      --zone <zone>              (aka -z)  zone.\n"
+		"      --keytag <keytag> | --cka_id <CKA_ID>      (aka -x | -k)\n"
+/*		"      [--force]                  (aka -f)  force even if there is no configured\n"
+		"                                           DelegationSignerSubmitCommand.\n" */
 	);
 }
 
@@ -67,47 +63,16 @@ static int
 run(int sockfd, engine_type* engine, const char *cmd, ssize_t n,
 	db_connection_t *dbconn)
 {
-	char buf[ODS_SE_MAXLINE];
-	const int NARGV = 8;
-	const char *argv[NARGV];
-	int argc;
-	task_type *task;
-	ods_status status;
-
-	ods_log_debug("[%s] %s command", module_str, key_ds_submit_funcblock()->cmdname);
-	/* consume command */
-	cmd = ods_check_command(cmd, n, key_ds_submit_funcblock()->cmdname);
-
-	// Use buf as an intermediate buffer for the command.
-	strncpy(buf, cmd, sizeof(buf));
-	buf[sizeof(buf)-1] = '\0';
-
-	// separate the arguments
-	argc = ods_str_explode(&buf[0], NARGV, &argv[0]);
-	if (argc > NARGV) {
-		ods_log_warning("[%s] too many arguments for %s command",
-						module_str, key_ds_submit_funcblock()->cmdname);
-		client_printf(sockfd,"too many arguments\n");
-		return -1;
+	int error;
+	/* TODO, this changes the state, but sbmt cmd is not exec. */
+	error = run_ds_cmd(sockfd, cmd, n, dbconn,
+		KEY_DATA_DS_AT_PARENT_SUBMIT,
+		KEY_DATA_DS_AT_PARENT_SUBMITTED);
+	if (error == 0) {
+		flush_enforce_task(engine, 0);
 	}
+	return error;
 
-	const char *zone = NULL;
-	const char *cka_id = NULL;
-	(void)ods_find_arg_and_param(&argc,argv,"zone","z",&zone);
-	(void)ods_find_arg_and_param(&argc,argv,"cka_id","k",&cka_id);
-	bool force = ods_find_arg(&argc,argv,"force","f") != -1;
-	bool bAutomatic = ods_find_arg(&argc,argv,"auto","a") != -1;
-	if (argc) {
-		ods_log_warning("[%s] unknown arguments for %s command",
-						module_str, key_ds_submit_funcblock()->cmdname);
-		client_printf(sockfd,"unknown arguments\n");
-		return -1;
-	}
-	
-	//TODO: Need more validation of the permitted command line options combinatio
-
-	/* perform task immediately */
-	return !perform_keystate_ds_submit(sockfd,engine->config,zone,cka_id,bAutomatic?1:0, force);
 }
 
 static struct cmd_func_block funcblock = {
